Separated classes into smaller packages Because: - It makes the code directory cleaner - It is good for Modularizing
This commit will separate classes into packages
